requirements.txt
dj-database-url==0.4.1
djangorestframework==3.4.4
python-decouple==3.0
